Use visualization also in main editor
