+++ Views/Visualization/GraphWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="JiME.Visualization.Views.GraphWindow"
+﻿<Window x:Class="JiME.Views.GraphWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:JiME.Visualization.Views"
+        xmlns:local="clr-namespace:JiME.Views"
         xmlns:controls="http://schemas.panthernet.ru/graphx/"
         mc:Ignorable="d"
         Title="GraphWindow" Height="450" Width="800" MinHeight="500" MinWidth="700" WindowStartupLocation="CenterScreen" Background="#FF2D2D30" ResizeMode="CanResizeWithGrip">
