--- Visualization/Views/GraphWindow.xaml.cs
@@ -1,140 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-using JiME.Visualization.Models;
-using GraphX.Common.Enums;
-using GraphX.Logic.Algorithms.LayoutAlgorithms;
-using GraphX.Controls;
-
-namespace JiME.Visualization.Views
-{
-    /// <summary>
-    /// Interaction logic for GraphWindow.xaml
-    /// </summary>
-    public partial class GraphWindow : Window
-    {
-        Scenario scenario;
-
-        public GraphWindow()
-        {
-            InitializeComponent();
-        }
-
-        public GraphWindow(Scenario s) : this()
-        {
-            scenario = s;
-
-            //Customize Zoombox a bit
-            //Set minimap (overview) window to be visible by default
-            ZoomControl.SetViewFinderVisibility(zoomctrl, Visibility.Visible);
-            //Set Fill zooming strategy so whole graph will be always visible
-            zoomctrl.ZoomToFill();
-
-            SetupGraph();
-
-            Loaded += MainWindow_Loaded;
-        }
-
-        void MainWindow_Loaded(object sender, RoutedEventArgs e)
-        {
-            //lets create graph
-            //Note that you can't create it in class constructor as there will be problems with visuals
-
-            //Lets generate configured graph using pre-created data graph assigned to LogicCore object.
-            //Optionaly we set first method param to True (True by default) so this method will automatically generate edges
-            //  If you want to increase performance in cases where edges don't need to be drawn at first you can set it to False.
-            //  You can also handle edge generation by calling manually Area.GenerateAllEdges() method.
-            //Optionaly we set second param to True (True by default) so this method will automaticaly checks and assigns missing unique data ids
-            //for edges and vertices in _dataGraph.
-            //Note! Area.Graph property will be replaced by supplied _dataGraph object (if any).
-            graphArea.GenerateGraph(true, true);
-
-            /* 
-             * After graph generation is finished you can apply some additional settings for newly created visual vertex and edge controls
-             * (VertexControl and EdgeControl classes).
-             * 
-             */
-
-            //This method sets the dash style for edges. It is applied to all edges in Area.EdgesList. You can also set dash property for
-            //each edge individually using EdgeControl.DashStyle property.
-            //For ex.: Area.EdgesList[0].DashStyle = GraphX.EdgeDashStyle.Dash;
-            graphArea.SetEdgesDashStyle(EdgeDashStyle.Solid);
-
-            //This method sets edges arrows visibility. It is also applied to all edges in Area.EdgesList. You can also set property for
-            //each edge individually using property, for ex: Area.EdgesList[0].ShowArrows = true;
-            graphArea.ShowAllEdgesArrows(true);
-
-            //This method sets edges labels visibility. It is also applied to all edges in Area.EdgesList. You can also set property for
-            //each edge individually using property, for ex: Area.EdgesList[0].ShowLabel = true;
-            graphArea.ShowAllEdgesLabels(true);
-
-            zoomctrl.ZoomToFill();
-        }
-
-        
-
-        private void SetupGraph()
-        {
-            if (scenario == null)
-            {
-                return;
-            }
-
-            // Setup the graph
-            var dataGraph = Graph.Generate(scenario);
-
-            // SETUP GRAPHAREA
-            //Lets create logic core and filled data graph with edges and vertices
-            var logicCore = new GraphLogic() { Graph = dataGraph };
-            //This property sets layout algorithm that will be used to calculate vertices positions
-            //Different algorithms uses different values and some of them uses edge Weight property.
-            logicCore.DefaultLayoutAlgorithm = LayoutAlgorithmTypeEnum.Tree;
-            //Now we can set parameters for selected algorithm using AlgorithmFactory property. This property provides methods for
-            //creating all available algorithms and algo parameters.
-            logicCore.DefaultLayoutAlgorithmParams = logicCore.AlgorithmFactory.CreateLayoutParameters(LayoutAlgorithmTypeEnum.Tree);
-            //Unfortunately to change algo parameters you need to specify params type which is different for every algorithm.
-            //((KKLayoutParameters)logicCore.DefaultLayoutAlgorithmParams).MaxIterations = 100;
-            ((SimpleTreeLayoutParameters)logicCore.DefaultLayoutAlgorithmParams).LayerGap = 100;
-            ((SimpleTreeLayoutParameters)logicCore.DefaultLayoutAlgorithmParams).VertexGap = 100;
-
-            //This property sets vertex overlap removal algorithm.
-            //Such algorithms help to arrange vertices in the layout so no one overlaps each other.
-            logicCore.DefaultOverlapRemovalAlgorithm = OverlapRemovalAlgorithmTypeEnum.FSA;
-            //Default parameters are created automaticaly when new default algorithm is set and previous params were NULL
-            logicCore.DefaultOverlapRemovalAlgorithmParams.HorizontalGap = 50;
-            logicCore.DefaultOverlapRemovalAlgorithmParams.VerticalGap = 50;
-
-            //This property sets edge routing algorithm that is used to build route paths according to algorithm logic.
-            //For ex., SimpleER algorithm will try to set edge paths around vertices so no edge will intersect any vertex.
-            //Bundling algorithm will try to tie different edges that follows same direction to a single channel making complex graphs more appealing.
-            logicCore.DefaultEdgeRoutingAlgorithm = EdgeRoutingAlgorithmTypeEnum.SimpleER;
-
-            // Add some separation on parallel edges
-            logicCore.EnableParallelEdges = true;
-            logicCore.ParallelEdgeDistance = 20;
-
-            //This property sets async algorithms computation so methods like: Area.RelayoutGraph() and Area.GenerateGraph()
-            //will run async with the UI thread. Completion of the specified methods can be catched by corresponding events:
-            //Area.RelayoutFinished and Area.GenerateGraphFinished.
-            logicCore.AsyncAlgorithmCompute = false;
-
-            //Finally assign logic core to GraphArea object
-            graphArea.LogicCore = logicCore;
-        }
-
-        private void SaveButton_Click(object sender, RoutedEventArgs e)
-        {
-            graphArea.ExportAsPng();
-        }
-    }
-}
